switch_env: suppress error

+++ newtool/pkgs/fish/functions/switch_env.fish
@@ -9,16 +9,20 @@ function __clear_gitconfig
 end
 
 function switch_env
-	switch $argv[1]
-	case "ckpd"
-		__clear_gitconfig
-		ln -s -f ~/.gitconfig.ckpd ~/.gitconfig
-		eval (ssh-agent -c)
-		echo "cookpad env"
-	case "priv"
-		__clear_gitconfig
-		ln -s -f ~/.gitconfig.priv ~/.gitconfig
-		set -gx SSH_AUTH_SOCK (gpgconf --list-dirs agent-ssh-socket)
-		echo "private env"
+	if test -e ~/.gitconfig.ckpd; and test -e ~/.gitconfig.priv
+		switch $argv[1]
+		case "ckpd"
+			__clear_gitconfig
+			ln -s -f ~/.gitconfig.ckpd ~/.gitconfig
+			eval (ssh-agent -c)
+			echo "cookpad env"
+		case "priv"
+			__clear_gitconfig
+			ln -s -f ~/.gitconfig.priv ~/.gitconfig
+			set -gx SSH_AUTH_SOCK (gpgconf --list-dirs agent-ssh-socket)
+			echo "private env"
+		end
+	else
+		echo "private only"
 	end
 end
